app/requirements.txt
amqp==5.1.1
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
asyncio==3.4.3
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.1
django-apscheduler==0.6.2
django-background-tasks==1.2.5
django-celery-beat==2.4.0
django-compat==1.0.15
django-environ==0.9.0
django-timezone-field==5.0
djangorestframework==3.13.1
environ==1.0
idna==3.3
install==1.3.5
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.31
psycopg2==2.9.5
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
requests==2.28.1
schedule==1.1.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
